refactor: move flattenSidebarTree to utils
